Release 0.11.7

See merge request devel/gamspy!197

+++ tests/test_gamspy.py
@@ -42,7 +42,7 @@ class GamspySuite(unittest.TestCase):
     def test_version(self):
         import gamspy
 
-        self.assertEqual(gamspy.__version__, "0.11.6")
+        self.assertEqual(gamspy.__version__, "0.11.7")
 
 
 class DocsSuite(unittest.TestCase):
@@ -206,7 +206,8 @@ def main():
 
     gams_paths = glob.glob("_*")
     for gams_path in gams_paths:
-        os.remove(gams_path)
+        if os.path.isfile(gams_path):
+            os.remove(gams_path)
 
     return 0
